MAINT: separate serving and formatting logic
